Better organization of Scene Changing scripts
