+++ Assets/Scripts/SceneChanging/ISceneChanger.cs
@@ -1,0 +1,7 @@
+﻿namespace SceneChanging
+{
+    public interface ISceneChanger
+    {
+        void SwitchTo(Scenes scene);
+    }
+}
