+++ Assets/Scripts/SceneChanging/SceneChangerAdditive.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace SceneChanging
+{
+    public class SceneChangerAdditive : ISceneChanger
+    {
+        public void SwitchTo(Scenes scene)
+            => Debug.LogWarning("Need to implement Additive Loading");
+    }
+}
