+++ Assets/Scripts/SceneChanging/SceneChangerDebugLog.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace SceneChanging
+{
+    public class SceneChangerDebugLog : ISceneChanger
+    {
+        public void SwitchTo(Scenes scene)
+            => Debug.Log(scene.ToString());
+    }
+}
