+++ Assets/Scripts/SceneChanging/SceneChangerSceneLoad.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine.SceneManagement;
+
+namespace SceneChanging
+{
+    public class SceneChangerSceneLoad : ISceneChanger
+    {
+        public void SwitchTo(Scenes scene)
+            => SceneManager.LoadScene(scene.ToString());
+    }
+}
