+++ Assets/Scripts/SceneChanging/Scenes.cs
@@ -1,0 +1,11 @@
+﻿namespace SceneChanging
+{
+    public enum Scenes
+    {
+        Original,
+        MegoPlayset,
+        LandOfGiants,
+        DancingOnCeiling,
+        NoGravity
+    }
+}
